rename and update test
